+++ api/routes/shop/[shopId]/resources/type/[resourceTypeId]/material/index.js
@@ -7,7 +7,7 @@ const materialSchema = z.object({
   title: z.string().min(1, "Material must have Title"),
   manufacturer: z.string().optional(),
   resourceTypeId: z.string().min(1, "Resource must have ID"),
-  costPerUnit: z.number().float().optional(),
+  costPerUnit: z.number().optional(),
   unitDescriptor: z.string().optional(),
   shopId: z.string().min(1, "Shop must have ID"),
 });
